Fix missing INSERT output for PDGITEM
Previously, the identifier column was not considered when deciding whether a row was "new". For PDGITEM, the only column other than the identifier (NAME) is ITEM_TYPE, which obviously isn't unique. In theory this could affect any table that has rows that agree on all considered columns other than the identifier. In practice, PDGITEM is the only one where it's likely. (And perhaps PDGITEM_MAP.)